chore: rename regex package to ruleset
The naming was easy to confuse with stdlib `regexp`.

Fixes #385
